+++ packages/boring-table-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # boring-table
 
+## 0.0.8
+
+### Patch Changes
+
+- feat plugin priorities
+  feat getRow on cells
+
+  add RowHidden
+  add new RowSelection
+
+  update filter
+
 ## 0.0.7
 
 ### Patch Changes